Data Binding

+++ BookWithMe/CoreData/ReviewEntity+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  ReviewEntity+CoreDataProperties.swift
 //  BookWithMe
 //
-//  Created by 계은성 on 5/14/25.
+//  Created by 계은성 on 5/15/25.
 //
 //
 
@@ -18,7 +18,7 @@ extension ReviewEntity {
 
     @NSManaged public var bookId: String?
     @NSManaged public var memorableQuotes: String?
-    @NSManaged public var rating: Int16
+    @NSManaged public var rating: Double
     @NSManaged public var reviewDetail: String?
     @NSManaged public var reviewSummary: String?
     @NSManaged public var tags: String?
